Changed Maritime so it updates the server correctly, tmp changes to ResourceBarController so it updates resources -could push later if we want
