+++ java/src/shared/model/GameManager.java
@@ -977,58 +977,6 @@ public class GameManager implements ModelSubject
 		return gameBank.getResourceCount(resourceType);
 	}
 	
-	/**
-	 * Takes resource from the bank and gives them to the current player
-	 * @param type
-	 * @param count
-	 * @throws ModelException
-	 */
-	public void giveResourcesToCurrentPlayer(ResourceType type, int count) throws ModelException{
-		System.out.println("In taking from bank: The bank has" + gameBank.getResourceCount(type) + " of " + type);
-		gameBank.getResource(type, count);
-		this.players.get(this.CurrentPlayersTurn()).playerBank.giveResource(type, count);
-		this.notifyCenter.notify(ModelNotification.RESOURCES);
-	}
-	
-	/**
-	 * Takes resources from the current player and gives them to the bank
-	 * @param type
-	 * @param count
-	 * @throws ModelException
-	 */
-	public void takeResourcesFromCurrentPlayer(ResourceType type, int count) throws ModelException{
-		gameBank.giveResource(type, count);
-		this.players.get(this.CurrentPlayersTurn()).playerBank.getResource(type, count);
-		this.notifyCenter.notify(ModelNotification.RESOURCES);
-	}
-	
-	
-	/**
-	 * Takes resources from the bank and gives them to the passed player index
-	 * @param playerIndex
-	 * @param type
-	 * @param count
-	 * @throws ModelException
-	 */
-	public void giveResourcesToPlayer(int playerIndex, ResourceType type, int count) throws ModelException{
-		gameBank.getResource(type, count);
-		this.players.get(playerIndex).playerBank.giveResource(type, count);
-		this.notifyCenter.notify(ModelNotification.RESOURCES);
-	}
-	
-	/**
-	 * Takes resources from the passed player index and gives them to the bank
-	 * @param playerIndex
-	 * @param type
-	 * @param count
-	 * @throws ModelException
-	 */
-	public void takeResourcesFromPlayer(int playerIndex, ResourceType type, int count) throws ModelException{
-		gameBank.giveResource(type, count);
-		this.players.get(playerIndex).playerBank.getResource(type, count);
-		this.notifyCenter.notify(ModelNotification.RESOURCES);
-	}
-	
 	public void initializeBankTempTesting(){
 		try {
 			gameBank.giveResource(ResourceType.BRICK, 60);
